refactor : 내부 클래스 분리 (#813)
